pybind/mgr: Remove code duplication
Introduced in #27512

(cherry picked from commit a19d166083fdb6d234a3464dbf06cd4f44a46774)

Conflicts:
	src/pybind/mgr/ansible/module.py
